requirements.txt
streamlit==1.43.1
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.3.2
pathlib==1.0.1
pillow==10.2.0